requirements.txt
python-multipart
fastapi==0.116.2
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.76
pydantic==2.11.9
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.36.0
ddgs==9.6.0
langchain-google-genai==2.0.10
supabase==2.10.0
pyjwt==2.10.1
email-validator==2.3.0
pydantic[email]
email-validator